reorganize launch files to isolate gazebo dependency (#133)
* reorganize launch files to isolate gazebo dependency


* move rviz files to bringup directory


* fix linter error line too long


* rename rviz variables since they are not indicating directories


* split irobot_create_bringup into common and gazebo, some minor other changes


* fix linter error
