+++ irobot_create_common_bringup/config/hazard_vector_params.yaml
