+++ irobot_create_description/CMakeLists.txt
@@ -3,14 +3,10 @@ cmake_minimum_required(VERSION 3.5)
 project(irobot_create_description)
 
 find_package(ament_cmake REQUIRED)
-find_package(urdf REQUIRED)
-find_package(xacro REQUIRED)
 
 install(
   DIRECTORY
-    launch
     meshes
-    rviz
     urdf
   DESTINATION
     share/${PROJECT_NAME}/
