+++ irobot_create_description/urdf/dock/standard_dock.urdf.xacro
