+++ irobot_create_gazebo_bringup/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>irobot_create_bringup</name>
+  <name>irobot_create_gazebo_bringup</name>
   <version>0.0.1</version>
-  <description>Provides launch and configuration scripts for a simulated iRobot(R) Create(R) 3 Educational Robot.</description>
+  <description>Provides launch and configuration scripts for a Gazebo simulated iRobot(R) Create(R) 3 Educational Robot.</description>
   <maintainer email="ekumen@irbt.onmicrosoft.com">Ekumen</maintainer>
   <license>BSD</license>
 
@@ -11,10 +11,9 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <depend>urdf</depend>
-  <depend>xacro</depend>
-
+  <exec_depend>gazebo_plugins</exec_depend>
   <exec_depend>gazebo_ros</exec_depend>
+  <exec_depend>irobot_create_common_bringup</exec_depend>
   <exec_depend>irobot_create_control</exec_depend>
   <exec_depend>irobot_create_description</exec_depend>
   <exec_depend>irobot_create_gazebo_plugins</exec_depend>
