+++ irobot_create_gazebo_sim/CMakeLists.txt
@@ -1,4 +1,4 @@
 cmake_minimum_required(VERSION 3.5)
-project(irobot_create_sim)
+project(irobot_create_gazebo_sim)
 find_package(ament_cmake REQUIRED)
 ament_package()
